terminando controller locacao

+++ Controller/Service/Models/LocacaoService.cs
@@ -1,7 +1,11 @@
 ﻿using AutoMapper;
+using Domain.DTO;
 using Domain.Models;
+using Microsoft.EntityFrameworkCore;
 using Repository.Models;
 using System;
+using System.Collections;
+using System.Collections.Generic;
 using System.Linq;
 
 namespace Service.Models
@@ -9,39 +13,80 @@ namespace Service.Models
     public class LocacaoService
     {
         private readonly LocacaoRepository _locacaoRepository;
+        private readonly FilmeService _filmeService;
+        private readonly ClienteService _clienteService;
         private readonly IMapper _mapper;
 
-        public LocacaoService(LocacaoRepository locacaoRepository, IMapper mapper)
+        public LocacaoService(LocacaoRepository locacaoRepository, FilmeService filmeService, ClienteService clienteService, IMapper mapper)
         {
             _locacaoRepository = locacaoRepository;
+            _filmeService = filmeService;
+            _clienteService = clienteService;
             _mapper = mapper;
         }
 
-        public object BuscarTodos()
+        public IList<LocacaoDTO> BuscarTodos()
         {
-            return _locacaoRepository.GetAll().ToList();
+            var list = _locacaoRepository
+                .GetAll()
+                .Include(x => x.Filme)
+                .Include(x => x.Cliente)
+                .ToList();
+
+            var listDTO = _mapper.Map<List<LocacaoDTO>>(list);
+
+            return listDTO;
         }
 
-        public object Alugar(Locacao locacao)
+        public Locacao Alugar(LocacaoAlugarDTO locacaoDTO)
         {
-            // verificar se o filme não está alugado
+            var filme = _filmeService.GetById(locacaoDTO.IdFilme);
+            if(filme == null)
+                throw new Exception("Filme não encontrado!");
 
+            var cliente = _clienteService.GetById(locacaoDTO.IdFilme);
+            if (cliente == null)
+                throw new Exception("Cliente não encontrado!");
+
+            var locacao = _mapper.Map<Locacao>(locacaoDTO);
             locacao.DataLocacao = DateTime.Now;
+
+            if(filme.Lancamento)
+                locacao.DataDevolucao = DateTime.Now.AddDays(2);
+            else
+                locacao.DataDevolucao = DateTime.Now.AddDays(3);
 
             _locacaoRepository.Save(locacao);
 
             return locacao;
         }
 
-        public object Devolver(Locacao locacao, int idCliente, int idFilme)
+        public Locacao Atualizar(LocacaoAtualizarDTO locacaoDTO, int idCliente, int idFilme)
         {
             _locacaoRepository.EncontrarLocacao(idCliente, idFilme);
 
-            // verificar se o filme já não foi devolvido
+            var locacao = _mapper.Map<Locacao>(locacaoDTO);
 
-            locacao.IdFilme = idFilme;
-            locacao.IdCliente = idCliente;
-            locacao.DataDevolucao = DateTime.Now;
+            _locacaoRepository.Update(locacao);
+
+            return locacao;
+        }
+
+        public Locacao Devolver(LocacaoAlugarDTO locacaoDTO)
+        {
+            var filme = _filmeService.GetById(locacaoDTO.IdFilme);
+            if (filme == null)
+                throw new Exception("Filme não encontrado!");
+
+            var cliente = _clienteService.GetById(locacaoDTO.IdFilme);
+            if (cliente == null)
+                throw new Exception("Cliente não encontrado!");
+
+            var locacao = _locacaoRepository.EncontrarLocacao(locacaoDTO.IdCliente, locacaoDTO.IdFilme);
+            if(locacao.FilmeDevolvido)
+                throw new Exception("Essa locação já foi devolvida!");
+
+            locacao.FilmeDevolvido = true;
 
             _locacaoRepository.Update(locacao);
 
@@ -50,9 +95,10 @@ namespace Service.Models
 
         public void Deletar(int idCliente, int idFilme)
         {
-            _locacaoRepository.EncontrarLocacao(idCliente, idFilme);
+            var locacao = _locacaoRepository.EncontrarLocacao(idCliente, idFilme);
 
-            // verificar se o filme não está alugado
+            if (!locacao.FilmeDevolvido)
+                throw new Exception("Esse Filme está alugado! Não é possivel excluir ele agora.");
 
             _locacaoRepository.Delete(x => x.IdCliente == idCliente && x.IdFilme == idFilme);
         }
